Move BaseModel, json_util, model_util to baselayer
