+++ cesium_app/models.py
@@ -9,33 +9,10 @@ from playhouse.postgres_ext import ArrayField, BinaryJSONField
 from playhouse.shortcuts import model_to_dict
 from playhouse import signals
 
-from cesium_app.json_util import to_json
+from baselayer.app.json_util import to_json
+from baselayer.app.models import BaseModel, db
+
 from cesium import featurize
-
-
-# The db has to be initialized later; this is done by the app itself
-# See `app_server.py`
-db = pw.PostgresqlDatabase(None, autocommit=True, autorollback=True)
-
-
-class BaseModel(signals.Model):
-    """All other models are derived from this one.
-
-    It adds the following:
-
-    - __dict__ method to convert the model to a dict
-    - __str__ method to convert the model to a JSON string: `str(my_model)`
-    - Specify the database in use
-
-    """
-    def __str__(self):
-        return to_json(self.__dict__())
-
-    def __dict__(self):
-        return model_to_dict(self, recurse=False, backrefs=False)
-
-    class Meta:
-        database = db
 
 
 class User(BaseModel):
